Split up Rhythm and RhythmLeaf
